pyproject.toml
[tool.poetry]
name = "nonebot-plugin-daily-oil-price"
version = "0.1.3"
description = "nonebot每日油价插件"
authors = ["wyeeeee <1563031828@qq.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = "^2.3.2"
nonebot-adapter-onebot = "^2.4.4"
nonebot-plugin-apscheduler = "^0.5.0"
httpx = "^0.27.0"
nonebot-plugin-localstore = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
